Split testUtils classes into individual files.
