+++ tests/distributed-transactions-remote-test.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 
-import testUtils
+from testUtils import Utils
+from Cluster import Cluster
+from TestHelper import TestHelper
 
-import argparse
 import subprocess
 import tempfile
 import os
@@ -13,32 +14,20 @@ import os
 #  distributed-transactions-remote-test.py -v --clean-run --dump-error-detail
 ###############################################################
 
-Print=testUtils.Utils.Print
+Print=Utils.Print
 
 def errorExit(msg="", errorCode=1):
     Print("ERROR:", msg)
     exit(errorCode)
 
-pnodes=1
-# nodesFile="tests/sample-cluster-map.json"
-parser = argparse.ArgumentParser()
-parser.add_argument("-p", type=int, help="producing nodes count", default=pnodes)
-parser.add_argument("-v", help="verbose", action='store_true')
-parser.add_argument("--leave-running", help="Leave cluster running after test finishes", action='store_true')
-parser.add_argument("--dump-error-details",
-                    help="Upon error print etc/eosio/node_*/config.ini and var/lib/node_*/stderr.log to stdout",
-                    action='store_true')
-parser.add_argument("--clean-run", help="Kill all nodeos and kleos instances", action='store_true')
-
-args = parser.parse_args()
+args = TestHelper.parse_args({"-p","--dump-error-details","-v","--leave-running","--clean-run"})
 pnodes=args.p
-# nodesFile=args.nodes_file
 debug=args.v
 dontKill=args.leave_running
 dumpErrorDetails=args.dump_error_details
 killAll=args.clean_run
 
-testUtils.Utils.Debug=debug
+Utils.Debug=debug
 
 killEosInstances=not dontKill
 topo="mesh"
@@ -61,7 +50,7 @@ clusterMapJsonTemplate="""{
 }
 """
 
-cluster=testUtils.Cluster()
+cluster=Cluster()
 
 (fd, nodesFile) = tempfile.mkstemp()
 try:
@@ -80,7 +69,7 @@ try:
     if not cluster.waitOnClusterBlockNumSync(3):
         errorExit("Cluster never stabilized")
 
-    producerKeys=testUtils.Cluster.parseClusterKeys(total_nodes)
+    producerKeys=Cluster.parseClusterKeys(total_nodes)
     defproduceraPrvtKey=producerKeys["defproducera"]["private"]
     defproducerbPrvtKey=producerKeys["defproducerb"]["private"]
 
@@ -100,13 +89,6 @@ try:
     Print("\nEND")
 finally:
     os.remove(nodesFile)
-    if not testSuccessful and dumpErrorDetails:
-        cluster.dumpErrorDetails()
-        Print("== Errors see above ==")
-
-    if killEosInstances:
-        Print("Shut down the cluster and cleanup.")
-        cluster.killall(allInstances=killAll)
-        cluster.cleanup()
+    TestHelper.shutdown(cluster, None, testSuccessful, killEosInstances, False, False, killAll, dumpErrorDetails)
 
 exit(0)
