+++ tests/restart-scenarios-test.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
 
-import testUtils
+from testUtils import Utils
+from Cluster import Cluster
+from WalletMgr import WalletMgr
 from TestHelper import TestHelper
 
 import random
@@ -22,7 +24,7 @@ import traceback
 ###############################################################
 
 
-Print=testUtils.Utils.Print
+Print=Utils.Print
 
 def errorExit(msg="", errorCode=1):
     Print("ERROR:", msg)
@@ -46,12 +48,12 @@ killAll=args.clean_run
 p2pPlugin=args.p2p_plugin
 
 seed=1
-testUtils.Utils.Debug=debug
+Utils.Debug=debug
 testSuccessful=False
 
 random.seed(seed) # Use a fixed seed for repeatability.
-cluster=testUtils.Cluster(walletd=True)
-walletMgr=testUtils.WalletMgr(True)
+cluster=Cluster(walletd=True)
+walletMgr=WalletMgr(True)
 
 try:
     cluster.setChainStrategy(chainSyncStrategyStr)
